user-deletion-consumer/requirements.txt
pika==1.3.2
PyMySQL==1.0.2
python-dotenv==0.21.0
